// src/app/services/models/proc-details.models.ts
export interface ProcDetailsModel {
    id: string;
    konkursId: string;
    konkursNr: string;
    konkursName: string;
    bukrs: string;
    orgKey: string;
    stat: string;
    crtDate: string;
    crtUser: string;
    lotId: string;
    lotNr: string;
    lotName: string;
    ekorg: string;
    finanLimitWovat: string;
    vatRate: string;
    finanLimitWvat: string;
    lifnr: string;
    lifnrName: string;
    ofrDate: string;
    ofrTime: string;
    priceNds: string;
    priceSNds: string;
    deliverDate: string;
    deliverTime: string;
    winFld: boolean
}
  